requirements.txt
Django==5.2.5
psycopg2-binary==2.9.9
pytest==8.3.2
pytest-django==4.9.0
pytest-cov==5.0.0
coverage==7.6.0
python-dotenv==1.0.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
watchdog==4.0.2
mypy==1.10.0
django-stubs==5.0.0
django-stubs-ext==5.0.0
